Comment out printed warning if not excluding any advected tracers
Config file GCHP.rc setting EXCLUDE_ADVECTION_TRACERS was recently added
to explicitly be set to NO. This avoids an ESMF error in advection code
triggered when using the default value of ALWAYS. However, the option NO
is not supported in GEOS and causes the line "Invalid option, ignored" to
be printed to the line every time FV3 is run. This update removes that
line. In a more recent version of FV3 there is handling for NO as a valid
option and this update is no longer necessary.